requirements.txt
Django==1.4.1
argparse==1.2.1
distribute==0.6.24
django-ratings==0.3.7
django-tinymce==1.5.1b2
fuzzywuzzy==0.1
wsgiref==0.1.2
